backend/requirements.txt
fastapi
uvicorn
sqlalchemy
firebase-admin
boto3
python-dotenv
python-docx
pdfkit
reportlab
requests 
psycopg2-binary 
html2text
beautifulsoup4
weasyprint
qrcode[pil]
Pillow
playwright
razorpay
python-multipart
python-jose[cryptography]
pydantic 